Fix changelogs:
As specified by @rhuk

Dates in the m/d/y or d-m-y formats are disambiguated by looking
at the separator between the various components: if the separator
is a slash (/), then the American m/d/y is assumed; whereas if the
separator is a dash (-) or a dot (.), then the European d-m-y
format is assumed. If, however, the year is given in a two digit
format and the separator is a dash (-, the date string is parsed
as y-m-d.